테스트 코드 수정
테스트 코드 수정

+++ Example-iOS/TestCollectionViewAdapter/AdapterDataTest/CountCollectionViewCell.swift
@@ -10,7 +10,7 @@ import UIKit
 import CollectionViewAdapter
 
 class CountCollectionViewCell: UICollectionViewCell, CollectionViewAdapterCellProtocol {
-    static var SpanSize: Int = 2
+    static var SpanSize: CGFloat = 2
     var actionClosure: ((_ name: String, _ object: Any?) -> Void)?
 
     @IBOutlet weak var label: UILabel!
